backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai==1.3.8
httpx==0.25.2
python-dotenv==1.0.0
neo4j==5.15.0
requests==2.31.0
